requirements.txt
python-telegram-bot
numpy
pandas
scikit-learn
matplotlib
tensorflow
tensorflow_addons
optuna
seaborn
xgboost
requests
fuzzywuzzy
